+++ src/com/zz/sdk/util/DebugFlags.java
@@ -92,8 +92,9 @@ public class DebugFlags {
 									+ "1、充值成功后，<font color='#800000'>一般1-10分钟即可到账</font>，简单方便。<br/>"
 									+ "2、充值卡充值请根据充值卡面额选择正确的充值金额，并仔细核对卡号和密码。<br/>"
 									+ "3、如有疑问请联系客服，客服热线：020-85525051 客服QQ：9159。");
-			s_debug_env.add(KeyUser.K_COIN_RATE, 0.01f);
-			s_debug_env.add(KeyUser.K_COIN_BALANCE, RANDOM.nextFloat() * 1987);
+			s_debug_env.add(KeyUser.K_COIN_RATE, Double.valueOf(0.01d));
+			s_debug_env.add(KeyUser.K_COIN_BALANCE,
+					Double.valueOf(RANDOM.nextDouble() * 1987));
 			return s_debug_env;
 		}
 		return env;
